fitness 1st

+++ src/main/java/com/fitness/myprojectBackend/service/HealthDataSer.java
@@ -5,7 +5,7 @@ import com.fitness.myprojectBackend.dto.HealthDataDto;
 import java.util.List;
 
 public interface HealthDataSer {
-    HealthDataDto submitHealthData(Long UserId,HealthDataDto healthDataDto);
+    HealthDataDto submitHealthData(int UserId,HealthDataDto healthDataDto);
     HealthDataDto getLatestHealthData(Long userId);
     List<HealthDataDto> getAllHealthData(Long userId);
 }
